Renamed database.dart to repository.dart
